back-end/app/requirements.txt
aiofiles==23.1.0
google.cloud==0.34.0
fastapi==0.99.1
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
googlemaps==4.10.0
grpcio==1.56.0
vertexai==0.0.1
starlette==0.27.0
google-cloud-aiplatform==1.27.1
uvicorn==0.22.0
